+++ packages/dumb/src/index.ts
@@ -61,9 +61,14 @@ import { RangeSelect } from './components/trends/cbg/range-select/range-select'
 import { TrendsProvider } from './provider/trends.provider'
 import { FocusedRangeLabels } from './components/trends/common/focused-range-labels/focused-range-labels'
 import { TrendsContainer } from './components/trends/common/trends-container/trends-container'
+import { SettingsPrintView } from './modules/print/settings-print-view/settings-print-view'
+import { PrintView } from './modules/print/print-view/print-view'
+import { LayoutColumnType } from './models/enums/layout-column-type.enum'
+import { buildLayoutColumns } from './modules/print/print-view/print-view.util'
 
 export {
   AverageDailyDoseStat,
+  buildLayoutColumns,
   type BgPrefs,
   BloodGlucoseTooltip,
   BolusTooltip,
@@ -78,12 +83,15 @@ export {
   FoodTooltip,
   HistoryParameterTable,
   formatBgValue,
+  LayoutColumnType,
   LoopModeStat,
   ParameterTooltip,
   PhysicalTooltip,
+  PrintView,
   PumpTable,
   RangeSelect,
   ReservoirTooltip,
+  SettingsPrintView,
   SimpleStat,
   StatTooltip,
   Table,