+++ packages/dumb/src/utils/datetime/datetime.util.ts
@@ -65,6 +65,14 @@ export const getSimpleHourFormatSpace = (): string => {
 }
 
 /**
+ * getLongDayFormat
+ * @returns string according to translation
+ */
+export const getLongDayFormat = (): string => {
+  return t('MMM D, YYYY')
+}
+
+/**
  * Moment format for:
  * - English: 'MMM D, YYYY h:mm a'
  * - French: 'D MMM YYYY, H:mm'