+++ packages/yourloops/app/app.tsx
@@ -29,8 +29,11 @@ import React from 'react'
 import { BrowserRouter } from 'react-router-dom'
 import { Auth0Provider } from '@auth0/auth0-react'
 
-import '@fontsource/roboto'
-import 'branding/theme.css'
+import '@fontsource/roboto/300.css'
+import '@fontsource/roboto/400.css'
+import '@fontsource/roboto/500.css'
+import '@fontsource/roboto/700.css'
+import 'branding/global.css'
 import 'classes.css'
 
 import appConfig from '../lib/config/config'