+++ packages/yourloops/components/buttons/icon-action.tsx
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021-2022, Diabeloop
+ * Copyright (c) 2021-2023, Diabeloop
  *
  * All rights reserved.
  *
@@ -30,7 +30,9 @@ import IconButton from '@mui/material/IconButton'
 import Tooltip from '@mui/material/Tooltip'
 
 interface IconActionButtonProps {
-  ariaLabel?: string
+  ['data-action']?: string
+  ['aria-label']?: string
+  ['data-testid']?: string
   className?: string
   color?: 'inherit' | 'primary' | 'secondary' | 'default'
   component?: React.ElementType
@@ -44,24 +46,24 @@ interface IconActionButtonProps {
 
 const IconActionButton: FunctionComponent<IconActionButtonProps> = (props) => {
   const { id, onClick, className, size, icon, disabled } = props
-  const color = props.color ?? 'primary'
+  const color = props.color
   const tooltip = props.tooltip ?? ''
 
   return (
-    <Tooltip title={tooltip} aria-label={tooltip} placement="bottom">
-      <div>
-        <IconButton
-          id={id}
-          size={size}
-          color={color}
-          onClick={onClick}
-          aria-label={props.ariaLabel}
-          className={className}
-          disabled={disabled}
-        >
-          {icon}
-        </IconButton>
-      </div>
+    <Tooltip title={tooltip} aria-label={tooltip}>
+      <IconButton
+        id={id}
+        size={size}
+        color={color}
+        onClick={onClick}
+        aria-label={props['aria-label']}
+        data-testid={props['data-testid']}
+        data-action={props['data-action']}
+        className={className}
+        disabled={disabled}
+      >
+        {icon}
+      </IconButton>
     </Tooltip>
   )
 }